Final/requirements.txt
Django==4.2
psycopg2-binary==2.9.3
pillow==11.0.0
djangorestframework
djangorestframework-simplejwt
drf-yasg